+++ lib/html2rss/feed_builder.rb
@@ -6,7 +6,7 @@ module Html2rss
     # Builds the requested feed type from the channel and article list.
     #
     # @param type [Symbol] :rss or :json_feed
-    # @param channel [Html2rss::Channel, Html2rss::Channel::Snapshot]
+    # @param channel [Html2rss::Channel]
     # @param articles [Array<Html2rss::Article>]
     # @return [RSS::Rss, Hash] format-compliant representation of the feed
     def self.build(type, channel:, articles:, **)